requirements.txt
discord.py >= 2.4
markovify
markdown
nio-bot[e2ee,cli]
portalocker
